타이머 시간이 제대로 작동하도록 하라
