+++ app-client/src/components/Pomodoro/Pomodoro.tsx
@@ -1,4 +1,3 @@
-import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import { TimerState } from '../../store/timer';
 
@@ -34,50 +33,11 @@ const BackContainer = styled.div`
 `;
 
 interface PomodoroProps {
-  time: number;
-  startTime: number;
+  percentage: number;
   timerState: TimerState;
 }
 
-function Pomodoro({ timerState, time, startTime }: PomodoroProps) {
-  const [percentage, setPercentage] = useState<number>(0);
-  const [animationFrameId, setAnimationFrameId] = useState<number | null>(null);
-
-  useEffect(() => {
-    if (timerState !== 'RUNNING') {
-      if (animationFrameId !== null) {
-        cancelAnimationFrame(animationFrameId);
-        setAnimationFrameId(null);
-        setPercentage(0);
-      }
-      return;
-    }
-
-    const animationDuration: number = time * 60 * 1000;
-
-    const animate = () => {
-      const currentTime = Date.now();
-      const elapsedTime = currentTime - startTime;
-
-      if (elapsedTime < animationDuration) {
-        setPercentage((elapsedTime / animationDuration) * 100);
-        setAnimationFrameId(requestAnimationFrame(animate));
-      } else {
-        setPercentage(100);
-      }
-    };
-
-    if (animationFrameId === null) {
-      setAnimationFrameId(requestAnimationFrame(animate));
-    }
-
-    return () => {
-      if (animationFrameId !== null) {
-        cancelAnimationFrame(animationFrameId);
-      }
-    };
-  }, [timerState, animationFrameId]);
-
+function Pomodoro({ timerState, percentage }: PomodoroProps) {
   return (
     <>
       {(timerState === 'INITIAL' || timerState === 'FINISHED') && (
