+++ app-client/src/components/Pomodoro/Timer.tsx
@@ -1,9 +1,5 @@
-import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 import { TimerState } from '../../store/timer';
-import { useTimerStore } from '../../store/timer';
-import { useTodosStore } from '../../store/todos';
-import { updatePerformPomodoro } from '../../services/todo';
 
 const Time = styled.div`
   color: var(--main-color);
@@ -15,46 +11,12 @@ const Time = styled.div`
 interface TimerProps {
   time: number;
   timerState: TimerState;
+  isBreak: boolean;
 }
 
-const Timer = ({ time, timerState }: TimerProps) => {
-  const USER_TIME = time * 60 * 1000;
-  const INTERVAL = 1000;
-  const [timerTime, setTimerTime] = useState<number>(USER_TIME);
-  const { complete } = useTimerStore();
-  const { todoId, selectedTodo, isTodoChange, setIsTodoChange } =
-    useTodosStore();
-
-  const minute = String(Math.floor((timerTime / (1000 * 60)) % 60)).padStart(
-    2,
-    '0',
-  );
-  const second = String(Math.floor((timerTime / 1000) % 60)).padStart(2, '0');
-
-  useEffect(() => {
-    if (timerState === 'INITIAL') {
-      setTimerTime(USER_TIME);
-    }
-
-    if (timerState === 'RUNNING') {
-      const timer = setInterval(() => {
-        setTimerTime((prevTime) => prevTime - INTERVAL);
-      }, INTERVAL);
-
-      if (timerTime <= 0) {
-        clearInterval(timer);
-        complete();
-        updatePerformPomodoro(todoId).then(() =>
-          setIsTodoChange(!isTodoChange),
-        );
-      }
-      return () => {
-        clearInterval(timer);
-      };
-    }
-  }, [time, timerState, timerTime]);
-
-  const isBreak: boolean = selectedTodo === '휴식';
+const Timer = ({ time, timerState, isBreak }: TimerProps) => {
+  const minute = String(Math.floor((time / (1000 * 60)) % 60)).padStart(2, '0');
+  const second = String(Math.floor((time / 1000) % 60)).padStart(2, '0');
 
   return (
     <>
